Fix wrong identifiers

+++ lib/thresholds.c
@@ -3,6 +3,8 @@
 #include "stdlib.h"
 #include "../plugins/utils.h"
 
+#include "./thresholds.h"
+
 char *range_to_string(range *input, enum value_type_t type) {
 	char *result = NULL;
 	if (input->alert_on == INSIDE) {
@@ -13,12 +15,12 @@ char *range_to_string(range *input, enum value_type_t type) {
 		xasprintf(&result, "~:");
 	} else  {
 		switch (type) {
-			case INT:
+			case INT64:
 				if (input->start.pd_int != 0) {
 					xasprintf(&result, "%i:", input->start.pd_int);
 				}
 				break;
-			case UINT:
+			case UINT64:
 				if (input->start.pd_uint != 0) {
 					xasprintf(&result, "%u:", input->start.pd_uint);
 				}
@@ -35,10 +37,10 @@ char *range_to_string(range *input, enum value_type_t type) {
 
 	if (!input->end_infinity) {
 		switch (type) {
-			case INT:
+			case INT64:
 				xasprintf(&result, "%i:", input->end.pd_int);
 				break;
-			case UINT:
+			case UINT64:
 				xasprintf(&result, "%u:", input->end.pd_uint);
 				break;
 			case DOUBLE:
